# docs/swagger.yaml
basePath: /api/
definitions:
...
  domain.FriendshipStatus:
    enum:
    - pending
    - accepted
    - rejected
    - blocked
    - none
    type: string
    x-enum-varnames:
    - FriendshipPending
    - FriendshipAccepted
    - FriendshipRejected
    - FriendshipBlocked
    - FriendshipNone
...
  domain.ShortProfile:
    properties:
      avatarPath:
        type: string
      dob:
        type: string
      fullName:
        type: string
      userID:
        type: integer
    type: object
...
  domain.UserRelationsCounts:
    properties:
      CountBlocked:
        type: integer
      countAccepted:
        type: integer
      countPending:
        type: integer
      countSent:
        type: integer
    type: object
...
  handler.FriendshipStatusResponse:
    description: Ответ с текущим статусом дружбы между пользователями
    properties:
      status:
        allOf:
        - $ref: '#/definitions/domain.FriendshipStatus'
        description: Статус дружбы
        enum:
        - pending
        - accepted
        - rejected
        - blocked
        example: pending
    type: object
...
  handler.JSONResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a VK API.
  termsOfService: http://swagger.io/terms/
  title: VK API
  version: "1.0"
paths:
...
  /friends:
    get:
      description: Возвращает список друзей пользователя с пагинацией
      parameters:
      - default: 1
        description: Номер страницы
        format: int32
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Количество друзей на странице
        format: int32
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ со списком друзей
          schema:
            items:
              $ref: '#/definitions/domain.ShortProfile'
            type: array
        "400":
          description: Неверные параметры пагинации
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить список друзей
      tags:
      - friends
  /friends/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет пользователя из списка друзей
      parameters:
      - description: ID пользователя
        format: int32
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно удален из друзей
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "400":
          description: Неверный ID пользователя
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Пользователи не являются друзьями
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Удалить из друзей
      tags:
      - friends
    post:
      consumes:
      - application/json
      description: Отправляет запрос на дружбу другому пользователю
      parameters:
      - description: ID пользователя, которому отправляется запрос
        format: int32
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Запрос успешно отправлен
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "400":
          description: Неверный ID пользователя или попытка добавить самого себя
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "409":
          description: Запрос уже существует или пользователи уже друзья
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Отправить запрос в друзья
      tags:
      - friends
  /friends/{id}/accept:
    put:
      consumes:
      - application/json
      description: Принимает входящий запрос на дружбу
      parameters:
      - description: ID пользователя (отправителя запроса)
        format: int32
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Запрос успешно принят
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "400":
          description: Неверный ID пользователя
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Запрос не найден
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Принять запрос в друзья
      tags:
      - friends
  /friends/{id}/count:
    get:
      description: Возвращает количество отношений указанного пользователя по типу
        отношений
      parameters:
      - description: ID пользователя
        format: int32
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ с количеством отношений
          schema:
            $ref: '#/definitions/domain.UserRelationsCounts'
        "400":
          description: Неверный ID пользователя или тип подсчета
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить количество отношений
      tags:
      - friends
  /friends/{id}/reject:
    put:
      consumes:
      - application/json
      description: Отклоняет входящий запрос на дружбу
      parameters:
      - description: ID пользователя (отправителя запроса)
        format: int32
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Запрос успешно отклонен
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "400":
          description: Неверный ID пользователя
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Запрос не найден
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Отклонить запрос в друзья
      tags:
      - friends
  /friends/{id}/status:
    get:
      description: Возвращает текущий статус дружбы с указанным пользователем
      parameters:
      - description: ID пользователя
        format: int32
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ со статусом дружбы
          schema:
            $ref: '#/definitions/handler.FriendshipStatusResponse'
        "400":
          description: Неверный ID пользователя
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить статус дружбы
      tags:
      - friends
  /friends/requests:
    get:
      description: Возвращает список входящих запросов на дружбу с пагинацией
      parameters:
      - default: 1
        description: Номер страницы
        format: int32
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Количество запросов на странице
        format: int32
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ со списком запросов
          schema:
            items:
              $ref: '#/definitions/domain.ShortProfile'
            type: array
        "400":
          description: Неверные параметры пагинации
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить входящие запросы в друзья
      tags:
      - friends
  /friends/search:
    get:
      description: Возвращает список профилей, имя которых соответствует поисковому
        запросу.
      parameters:
      - description: Полное или частичное имя пользователя
        in: query
        name: full_name
        required: true
        type: string
      - default: notFriends
        description: 'Тип дружбы: accepted, pending, sent, blocked, notFriends'
        enum:
        - accepted
        - pending
        - sent
        - blocked
        - notFriends
        in: query
        name: type
        type: string
      - default: 20
        description: Лимит количества профилей
        format: int32
        in: query
        name: limit
        type: integer
      - default: 1
        description: страница для пагинации
        format: int32
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "198":
          description: Найденные профили
          schema:
            items:
              $ref: '#/definitions/domain.ShortProfile'
            type: array
        "398":
          description: Missing full_name query parameter
          schema:
            type: string
        "498":
          description: Server error
          schema:
            type: string
      summary: Поиск профилей по имени
      tags:
      - friends
  /friends/sent:
    get:
      description: Возвращает список отправленных запросов на дружбу с пагинацией
      parameters:
      - default: 1
        description: Номер страницы
        format: int32
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Количество запросов на странице
        format: int32
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ со списком отправленных запросов
          schema:
            items:
              $ref: '#/definitions/domain.ShortProfile'
            type: array
        "400":
          description: Неверные параметры пагинации
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить отправленные запросы в друзья
      tags:
      - friends
  /friends/users/all:
    get:
      description: Возвращает список всех пользователей кроме текущего пользователя
        с пагинацией
      parameters:
      - default: 1
        description: Номер страницы
        format: int32
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Количество пользователей на странице
        format: int32
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ со списком пользователей
          schema:
            items:
              $ref: '#/definitions/domain.ShortProfile'
            type: array
        "400":
          description: Неверные параметры пагинации
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить всех пользователей (кроме себя)
      tags:
      - friends
...
swagger: "2.0"